Fix variable type to match query return type.
No idea why this is suddenly flagged now, though it is correct.

+++ src/server/wesnothd/player.cpp
@@ -15,7 +15,7 @@
 
 #include "server/wesnothd/player.hpp"
 
-wesnothd::player::player(const std::string& n, simple_wml::node& cfg, int id,
+wesnothd::player::player(const std::string& n, simple_wml::node& cfg, long id,
                          bool registered, const std::string& version, const std::string& source, unsigned long long login_id, const std::size_t max_messages,
                          const std::size_t time_period,
                          const bool moderator)
